nursing_final_build/requirements.txt
streamlit
openai
langchain-openai
langchain-community
chromadb
pypdf
python-dotenv
tiktoken
pysqlite3-binary
